api with call side rendering
