+++ oldwork-routing/src/app/study/[...lecture]/page.js
