+++ src/app/page.js
@@ -1,36 +1,24 @@
 "use client"
-import { useState } from 'react'
-import styles from './page.module.css'
+import "./globals.css"
 import Link from 'next/link'
-import { useRouter } from 'next/navigation'
+
 
 export default function Home() {
-  const router=useRouter()
 
-  const navigate=(name)=>{
+    return (
+        <main >
+            
+            <h1>Fetch Data with API in Client Component</h1>
+            <ul className="navbar">
+                <li>
+                   <Link href="/productlist">Go to Product List Page with client side render</Link>
+                </li>
+                <li>
+                   <Link href="/listproduct">Go to Product Lists Page with server render </Link>
+                </li>
+            </ul>
 
-    router.push(name)
-    
-  }
-
-  return (
-    <main >
-     <h1>Basic Routing | Make New Page </h1>
-     <br></br>
-     <Link href="/login">Go to Login Page</Link>
-     <br></br>
-     <br></br>
-     <Link href="/about">Go to About Page</Link>
-     <br></br>
-     <br></br>
-     <button onClick={()=>navigate("/login")}>Go to Login Page</button>
-     <br></br>
-     <br></br>
-     <button onClick={()=>navigate("/about")}>Go to About Page</button>
-     <br></br>
-     <br></br>
-     <button onClick={()=>navigate("/studentlist")}>Go to Student List Page</button>
-    </main>
-  )
+        </main>
+    )
 }
